Improve handling of empty pool_Pools table when pool__Members are present
Problem: When a user does not provide any pools in the pool__Pools table
but DOES provide pool__Members the previous behaviour was not as
user-friendly as possible.  Pool creation would be skipped due to any
empty pool__Pools table resulting in successful deployment without pool
creation.

Analysis: Update implementation layer to detect when pool__Members is
non-empty but pool__Pools is empty.  When this condition is detected
populate the pool__Pools table with default values based on the Indexes
found in the pool__Members table.  Additionally if a monitor with
Index 0 exists in the monitor__Monitors table use that monitor for the
auto-created pools.

Tests: Add test_pools_fixup.json test to the test framework.  Ran all
pool tests against test harness to verify no breakage.